[re-factory] add hooks to handle cursor
